Success caching PostDto(Post, Comment, Reaction) using Redis

+++ src/main/java/com/finalterm/alumninetwork/service/impl/CommentServiceImpl.java
@@ -1,13 +1,16 @@
 package com.finalterm.alumninetwork.service.impl;
 
 import com.finalterm.alumninetwork.dto.response.CommentDto;
+import com.finalterm.alumninetwork.dto.response.PostDTO;
 import com.finalterm.alumninetwork.mapper.CommentMapper;
+import com.finalterm.alumninetwork.mapper.PostMapper;
 import com.finalterm.alumninetwork.pojo.Comment;
 import com.finalterm.alumninetwork.pojo.Post;
 import com.finalterm.alumninetwork.pojo.User;
 import com.finalterm.alumninetwork.repository.CommentRepository;
 import com.finalterm.alumninetwork.service.CommentService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,6 +18,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 @Service
@@ -22,6 +26,9 @@ public class CommentServiceImpl implements CommentService {
 
     @Autowired
     private CommentRepository commentRepository;
+
+    @Autowired
+    private RedisTemplate redisTemplate;
 
     @Override
     @Transactional
@@ -45,11 +52,23 @@ public class CommentServiceImpl implements CommentService {
                 //Update comment parent
                 parentComment.getReplies().add(savedComment);
                 this.commentRepository.saveOrUpdate(parentComment);
+
+                String commentCountKey = "post:" + post.getId() + ":commentCount";
+                redisTemplate.opsForValue().increment(commentCountKey);
+
+                //Xoa cache cu de sau khi them comment moi
+                redisTemplate.delete("post:" + savedComment.getPost().getId() + ":comments");
+
                 return CommentMapper.toCommentDTO(savedComment);
             } else throw new RuntimeException("parentCommentId is null");
         } else {
             comment.setParentCommentId(null);
             this.commentRepository.saveOrUpdate(comment);
+
+            String commentCountKey = "post:" + post.getId() + ":commentCount";
+            redisTemplate.opsForValue().increment(commentCountKey);
+            //Xoa cache cu de sau khi them comment moi
+            redisTemplate.delete("post:" + comment.getPost().getId() + ":comments");
             return CommentMapper.toCommentDTO(comment);
         }
     }
@@ -57,19 +76,57 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional
     public List<CommentDto> getRootCommentsByPostId(int postId) {
-        return this.commentRepository.getRootCommentsByPostId(postId).stream().map(CommentMapper::toCommentDTO).collect(Collectors.toList());
+        String key = "post:" + postId + ":comments";
+        Object cached = redisTemplate.opsForValue().get(key);
+
+        if (cached != null) {
+            return (List<CommentDto>) cached;
+        }
+
+        List<CommentDto> listRootComments = this.commentRepository.getRootCommentsByPostId(postId)
+                .stream()
+                .map(CommentMapper::toCommentDTO)
+                .collect(Collectors.toList());
+
+        redisTemplate.opsForValue().set(key, listRootComments, 5, TimeUnit.MINUTES);
+        return listRootComments;
     }
 
     @Override
     @Transactional
     public List<CommentDto> getCommentsByParentCommentId(int parentCommentId) {
-        return this.commentRepository.getCommentsByParentCommentId(parentCommentId).stream().map(CommentMapper::toCommentDTO).collect(Collectors.toList());
+        String key = "parentComments:" + parentCommentId + ":comments";
+        Object cached = redisTemplate.opsForValue().get(key);
+
+        if (cached != null) {
+            return (List<CommentDto>) cached;
+        }
+
+        List<CommentDto> listChildComments = this.commentRepository.getCommentsByParentCommentId(parentCommentId)
+                .stream()
+                .map(CommentMapper::toCommentDTO)
+                .collect(Collectors.toList());
+
+        redisTemplate.opsForValue().set(key, listChildComments, 5, TimeUnit.MINUTES);
+        return listChildComments;
     }
 
     @Override
     @Transactional
     public void deleteComment(int commentId) {
-        this.commentRepository.deleteComment(commentId);
+        Comment comment = this.commentRepository.getCommentById(commentId);
+        if (comment != null) {
+            int postId = comment.getPost().getId();
+            this.commentRepository.deleteComment(commentId);
+            String commentCountKey = "post:" + postId + ":commentCount";
+
+            Integer count = redisTemplate.opsForValue().decrement(commentCountKey).intValue();
+
+            if (count < 0) {
+                redisTemplate.opsForValue().set(commentCountKey, 0);
+            }
+            redisTemplate.delete("post:" + postId + ":comments");
+        }
     }
 
     @Transactional
@@ -82,5 +139,16 @@ public class CommentServiceImpl implements CommentService {
     @Transactional
     public CommentDto updateComment(Comment comment) {
         return CommentMapper.toCommentDTO(this.commentRepository.saveOrUpdate(comment));
+    }
+
+    @Override
+    public Integer getTotalCommentByPostId(int postId) {
+        String key = "post:" + postId + ":commentCount";
+        Integer count = (Integer) redisTemplate.opsForValue().get(key);
+        if (count == null) {
+            count = commentRepository.countTotalCommentsByPostId(postId);
+            redisTemplate.opsForValue().set(key, count, 10, TimeUnit.MINUTES);
+        }
+        return count;
     }
 }
